requirements.txt
# This is a requirements file for a Python project.
pandas 
openpyxl 
whisper 
torch 
torchvision 
torchaudio 
transformers 
requests
openai
pydub
python-dotenv
